Rename services folder to core
