+++ services/matching-service/requirements.txt
@@ -17,3 +17,4 @@ uvicorn==0.37.0
 uvloop==0.21.0
 watchfiles==1.1.0
 websockets==15.0.1
+redis==6.4.0
